chore: exported types to integrations folder
