+++ packages/workspace-fusion/src/lib/integrations/README.md
@@ -1,0 +1,4 @@
+# Integrations folder
+
+This folder holds all the types for the integrations.
+Beware that any changes to this folder and its files will have a direct impact on the end user.
