+++ packages/workspace-fusion/src/lib/types/configuration.ts
@@ -1,22 +1,9 @@
-import { StatusItem } from '../integrations/status-bar';
 import { WorkspaceOnClick } from './onClick';
 
 export type SidesheetConfig<TData> = {
 	Component: (props: WorkspaceOnClick<TData>) => JSX.Element;
 	getTitle: (clickEv: WorkspaceOnClick<TData>) => string;
 };
-
-export type DataSourceOptions<TData> = {
-	/** Function for getting response object from server */
-	getResponseAsync: (signal?: AbortSignal) => Promise<Response>;
-	/**
-	 * Function for parsing response
-	 * Can be omitted if all you do is .json();
-	 */
-	responseParser?: (response: Response) => TData[] | Promise<TData[]>;
-};
-
-export type StatusBarConfig<TData> = (data: TData[]) => [StatusItem, ...StatusItem[]];
 
 export type CustomTabProps<TData> = {
 	data: TData[];
@@ -41,23 +28,6 @@ export type AppConfig<TabNames extends string> = {
 	appKey: string;
 	appColor: string;
 	defaultTab: TabNames;
-};
-
-export type PowerBiConfig = {
-	reportUri: string;
-	getConfig: (reportUri: string) => Promise<any>;
-	getToken: (reportUri: string) => Promise<PowerBiToken>;
-};
-/**
- * Configuration for adding a fusion power bi client
- * Requires client configured for fusion reports api
- */
-export type FusionPowerBiConfig = {
-	reportUri: string;
-};
-
-type PowerBiToken = {
-	token: string;
 };
 
 /** Any http client, with fusion scope */
